requirements.txt
firebase-admin
firebase-functions
google-genai
python-dotenv